add a few comments/style
